board object
